+++ Sources/CodexBar/SettingsStore+Defaults.swift
@@ -964,6 +964,30 @@ extension SettingsStore {
         }
     }
 
+    var mergeIconsStacked: Bool {
+        get { self.defaultsState.mergeIconsStacked }
+        set {
+            self.defaultsState.mergeIconsStacked = newValue
+            self.userDefaults.set(newValue, forKey: "mergeIconsStacked")
+        }
+    }
+
+    var mergeIconStackedTopProviderRaw: String? {
+        get { self.defaultsState.mergeIconStackedTopProviderRaw }
+        set {
+            self.defaultsState.mergeIconStackedTopProviderRaw = newValue
+            self.userDefaults.set(newValue, forKey: "mergeIconStackedTopProvider")
+        }
+    }
+
+    var mergeIconStackedBottomProviderRaw: String? {
+        get { self.defaultsState.mergeIconStackedBottomProviderRaw }
+        set {
+            self.defaultsState.mergeIconStackedBottomProviderRaw = newValue
+            self.userDefaults.set(newValue, forKey: "mergeIconStackedBottomProvider")
+        }
+    }
+
     var mergedMenuLastSelectedWasOverview: Bool {
         get { self.mergedMenuLastSelectedWasOverviewStorage }
         set {