+++ Sources/CodexBar/SettingsStore+MenuObservation.swift
@@ -102,6 +102,9 @@ extension SettingsStore {
         _ = self.mergeIcons
         _ = self.mergedOverviewLayout
         _ = self.switcherShowsIcons
+        _ = self.mergeIconsStacked
+        _ = self.mergeIconStackedTopProviderRaw
+        _ = self.mergeIconStackedBottomProviderRaw
         _ = self.mergedOverviewSelectedProviders
         _ = self.zaiAPIToken
         _ = self[providerConfig: .synthetic, field: .apiKey]